refactor(responses): clean up OpenAI agent loop handoff
