+++ model_gateway/benches/routing_allocation_bench.rs
@@ -68,7 +68,10 @@ fn extract_text_for_routing_old(req: &ResponsesRequest) -> String {
                 ResponseInputOutputItem::FunctionCallOutput { output, .. } => Some(output.clone()),
                 ResponseInputOutputItem::McpApprovalRequest { .. } => None,
                 ResponseInputOutputItem::McpApprovalResponse { .. } => None,
-                ResponseInputOutputItem::ImageGenerationCall { .. } => None,
+                ResponseInputOutputItem::WebSearchCall { .. }
+                | ResponseInputOutputItem::CodeInterpreterCall { .. }
+                | ResponseInputOutputItem::FileSearchCall { .. }
+                | ResponseInputOutputItem::ImageGenerationCall { .. } => None,
                 ResponseInputOutputItem::Compaction { .. } => None,
                 ResponseInputOutputItem::ComputerCall { .. } => None,
                 ResponseInputOutputItem::ComputerCallOutput { .. } => None,
