+++ model_gateway/src/routers/openai/responses/mod.rs
@@ -2,10 +2,10 @@
 //!
 //! This module provides comprehensive support for OpenAI Responses API with:
 //! - Streaming and non-streaming response handling
-//! - MCP (Model Context Protocol) tool interception and execution
+//! - MCP (Model Context Protocol) tool preparation and common-loop handoff
 //! - SSE (Server-Sent Events) parsing and forwarding
 //! - Response accumulation for persistence
-//! - Tool call detection and output index remapping
+//! - Upstream tool-call parsing and output index remapping
 //! - Input history loading from conversations and response chains
 //! - Shared helpers for response retrieval-related logic
 
@@ -16,9 +16,9 @@ pub(crate) mod history;
 mod non_streaming;
 pub(crate) mod route;
 mod streaming;
+mod upstream_stream_parser;
 mod utils;
 
-// Re-exported for openai::mcp::tool_handler (cross-module dependency)
 pub(crate) use accumulator::StreamingResponseAccumulator;
 pub(crate) use common::{extract_output_index, get_event_type};
 pub use non_streaming::handle_non_streaming_response;
